# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trigger/{kb_name}/": {
      "post": {
        "tags": [
          "trigger"
        ],
        "summary": "Create Trigger",
        "description": "Create a new trigger for a knowledge base.\n\nArgs:\n    kb_name: Knowledge base name\n    request: Trigger creation request\n\nReturns:\n    APIResponse with created trigger\n\nRaises:\n    HTTPException: If validation fails",
        "operationId": "create_trigger_api_v1_trigger__kb_name___post",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "trigger"
        ],
        "summary": "Get Triggers",
        "description": "Get all triggers for a knowledge base.\n\nArgs:\n    kb_name: Knowledge base name\n\nReturns:\n    APIResponse with list of triggers",
        "operationId": "get_triggers_api_v1_trigger__kb_name___get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trigger/{kb_name}/{trigger_id}": {
      "get": {
        "tags": [
          "trigger"
        ],
        "summary": "Get Trigger",
        "description": "Get a specific trigger by ID.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n\nReturns:\n    APIResponse with trigger details\n\nRaises:\n    HTTPException: If trigger not found",
        "operationId": "get_trigger_api_v1_trigger__kb_name___trigger_id__get",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "trigger"
        ],
        "summary": "Update Trigger",
        "description": "Update an existing trigger.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n    request: Update request (partial fields)\n\nReturns:\n    APIResponse with updated trigger\n\nRaises:\n    HTTPException: If validation fails",
        "operationId": "update_trigger_api_v1_trigger__kb_name___trigger_id__put",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trigger"
        ],
        "summary": "Delete Trigger",
        "description": "Delete a trigger from a knowledge base.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n\nReturns:\n    APIResponse confirming deletion\n\nRaises:\n    HTTPException: If KB not found",
        "operationId": "delete_trigger_api_v1_trigger__kb_name___trigger_id__delete",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trigger/{kb_name}/{trigger_id}/enable": {
      "patch": {
        "tags": [
          "trigger"
        ],
        "summary": "Enable Trigger",
        "description": "Enable a trigger.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n\nReturns:\n    APIResponse with updated trigger\n\nRaises:\n    HTTPException: If validation fails",
        "operationId": "enable_trigger_api_v1_trigger__kb_name___trigger_id__enable_patch",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trigger/{kb_name}/{trigger_id}/disable": {
      "patch": {
        "tags": [
          "trigger"
        ],
        "summary": "Disable Trigger",
        "description": "Disable a trigger.\n\nArgs:\n    kb_name: Knowledge base name\n    trigger_id: Trigger ID\n\nReturns:\n    APIResponse with updated trigger\n\nRaises:\n    HTTPException: If validation fails",
        "operationId": "disable_trigger_api_v1_trigger__kb_name___trigger_id__disable_patch",
        "parameters": [
          {
            "name": "kb_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Name"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trigger Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {

              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TriggerCondition": {
        "properties": {
          "field_name": {
            "type": "string",
            "title": "Field Name",
            "description": "Field name to match (e.g., 'sys_status')"
          },
          "field_value": {
            "type": "string",
            "title": "Field Value",
            "description": "Expected field value"
          },
          "operator": {
            "type": "string",
            "enum": [
              "eq",
              "ne",
              "contains",
              "startswith",
              "endswith"
            ],
            "title": "Operator",
            "description": "Comparison operator: eq=equals, ne=not equals, contains=contains substring",
            "default": "eq"
          }
        },
        "type": "object",
        "required": [
          "field_name",
          "field_value"
        ],
        "title": "TriggerCondition",
        "description": "Trigger condition for matching knowledge base data.\n\nConditions use AND logic - all conditions must be satisfied for trigger to fire."
      },
      "TriggerCreateRequest": {
        "properties": {
          "trigger_name": {
            "type": "string",
            "title": "Trigger Name",
            "description": "Human-readable trigger name"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Callback URL endpoint"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/TriggerCondition"
            },
            "type": "array",
            "minItems": 1,
            "title": "Conditions",
            "description": "List of conditions (AND logic)"
          },
          "http_method": {
            "type": "string",
            "enum": [
              "POST",
              "PUT",
              "PATCH"
            ],
            "title": "Http Method",
            "default": "POST"
          },
          "http_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Http Headers"
          },
          "batch_mode": {
            "type": "boolean",
            "title": "Batch Mode",
            "default": false
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size",
            "default": 50
          },
          "timeout": {
            "type": "integer",
            "maximum": 300,
            "minimum": 1,
            "title": "Timeout",
            "default": 30
          },
          "retry_times": {
            "type": "integer",
            "maximum": 10,
            "minimum": 0,
            "title": "Retry Times",
            "default": 3
          },
          "retry_interval": {
            "type": "integer",
            "maximum": 60,
            "minimum": 1,
            "title": "Retry Interval",
            "default": 5
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "update_data_enabled": {
            "type": "boolean",
            "title": "Update Data Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "trigger_name",
          "url",
          "conditions"
        ],
        "title": "TriggerCreateRequest",
        "description": "Request model for creating a new trigger."
      },
      "TriggerUpdateRequest": {
        "properties": {
          "trigger_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TriggerCondition"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "http_method": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "POST",
                  "PUT",
                  "PATCH"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Method"
          },
          "http_headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Headers"
          },
          "batch_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Mode"
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout"
          },
          "retry_times": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Times"
          },
          "retry_interval": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Interval"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "update_data_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Data Enabled"
          }
        },
        "type": "object",
        "title": "TriggerUpdateRequest",
        "description": "Request model for updating an existing trigger.\n\nAll fields are optional - only provided fields will be updated."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}